+++ source/extensions/omni.isaac.core/omni/isaac/core/tasks/pick_place.py
@@ -39,7 +39,6 @@ class PickPlace(BaseTask):
         self._cube_size = cube_size
         if self._cube_size is None:
             self._cube_size = np.array([0.0515, 0.0515, 0.0515]) / get_stage_units()
-        self._offset = offset
         if self._cube_initial_position is None:
             self._cube_initial_position = np.array([0.3, 0.3, 0.3]) / get_stage_units()
         if self._cube_initial_orientation is None: